Move payment tests to integration tree
